refactor: Activity 클래스들을 ui 패키지로 이동
