+++ app/src/main/AndroidManifest.xml
@@ -13,17 +13,17 @@
         android:theme="@style/Theme.Shopping"
         tools:targetApi="31" >
         <activity
-            android:name=".CartActivity"
+            android:name=".ui.cart.activity.CartActivity"
             android:exported="false"
             android:label="@string/title_activity_cart"
             android:theme="@style/Theme.Shopping" />
         <activity
-            android:name=".ProductDetailActivity"
+            android:name=".ui.productdetail.activity.ProductDetailActivity"
             android:exported="false"
             android:label="@string/title_activity_product_detail"
             android:theme="@style/Theme.Shopping" />
         <activity
-            android:name=".ProductListActivity"
+            android:name=".ui.shopping.activity.ProductListActivity"
             android:exported="true" >
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
