+++ app/src/main/java/woowacourse/shopping/ui/cart/activity/CartActivity.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping
+package woowacourse.shopping.ui.cart.activity
 
 import android.os.Bundle
 import androidx.activity.ComponentActivity
